# src/openapi.yaml
openapi: 3.1.0
info:
  title: Branch Bank API
  version: 1.0.0
  description: |
    The branch bank API provides user-facing operations for account lifecycle management
    and fund transfers in the distributed banking system.

    This API allows users to register with a branch bank, create accounts with chosen
    currencies, perform unauthenticated account lookups to verify ownership, and initiate
    same-bank and cross-bank fund transfers.

    ## Account Number Semantics

    ### Account Number Format
    - Account numbers are exactly 8 characters long
    - The first 3 characters represent the bank identifier (matching the bankId from the central bank registry)
    - The last 5 characters are unique within the bank and generated by the branch bank
    - Format: `{BANK_PREFIX}{UNIQUE_SUFFIX}` where BANK_PREFIX is 3 uppercase letters/digits and UNIQUE_SUFFIX is 5 alphanumeric characters

    Example: "EST12345" where:
    - "EST" = bank prefix (e.g., Estonia Commercial Bank)
    - "12345" = unique account identifier within that bank

    ### Account Lookup Behavior
    - GET /accounts/{accountNumber} is an unauthenticated endpoint
    - Returns the owner's full name when the account exists
    - Returns 404 Not Found when the account does not exist
    - No authentication is required for lookup, making it possible to verify account ownership before transfers

    This contract is implementation-agnostic and serves as the source of truth for interoperability
...

  contact:
...
servers:
...
paths:
  /users:
    post:
      tags:
        - Users
      summary: Register a new user
      description: |
        Registers a new user with the branch bank. Users must be registered before they can create accounts.

        **Registration Flow:**
        - User provides their full name and optional email address
        - Branch bank assigns a unique userId
        - User can immediately create accounts after registration

        **User ID Format:**
        User IDs are unique within the branch bank and follow the pattern: `user-{UUID}` where UUID is a standard UUID v4 format.

        **Note:** This endpoint does not require authentication for initial registration, as it is the first step in creating a banking relationship.
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
            examples:
              userRegistration:
                summary: Standard user registration
                value:
                  fullName: "Jane Doe"
                  email: "jane.doe@example.com"
              minimalRegistration:
                summary: Minimal user registration (email optional)
                value:
                  fullName: "John Smith"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRegistrationResponse'
              examples:
                successfulRegistration:
                  summary: User registered
                  value:
                    userId: "user-550e8400-e29b-41d4-a716-446655440000"
                    fullName: "Jane Doe"
                    email: "jane.doe@example.com"
                    createdAt: "2026-04-02T11:15:00Z"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidRequest:
                  summary: Missing required fields
                  value:
                    code: "INVALID_REQUEST"
                    message: "Full name is required"
        '409':
          description: User already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                duplicateUser:
                  summary: User with same email already exists
                  value:
                    code: "DUPLICATE_USER"
                    message: "A user with this email address is already registered"
  /users/{userId}:
    get:
      tags:
        - Users
      summary: Get user profile
      description: Returns user information including their accounts.
      operationId: getUser
      security: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            example: "user-550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  fullName:
                    type: string
                  email:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
                  accounts:
                    type: array
                    items:
                      type: object
                      properties:
                        account_number:
                          type: string
                        currency:
                          type: string
                        balance:
                          type: string
                        created_at:
                          type: string
        '404':
          description: User not found
  /users/{userId}/accounts:
    get:
      tags:
        - Accounts
      summary: List user accounts with balances
      description: Returns all accounts owned by the user with current balances. Requires authentication.
      operationId: listUserAccounts
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Accounts listed
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts:
                    type: array
                    items:
                      type: object
                      properties:
                        account_number:
                          type: string
                        currency:
                          type: string
                        balance:
                          type: string
                        created_at:
                          type: string
        '401':
          description: Unauthorized
    post:
      tags:
        - Accounts
      summary: Create a new account for a user
      description: |
        Creates a new bank account for an existing user with the specified currency.

        **Account Creation Flow:**
        - User must be registered (valid userId required)
        - User specifies the desired currency for the account
        - Branch bank generates a unique 8-character account number
        - Account number format: `{BANK_PREFIX}{UNIQUE_SUFFIX}` (3 + 5 characters)
        - Initial balance is set to 0.00

        **Account Number Rules:**
        - Account numbers must be exactly 8 characters
        - The first 3 characters must match the bank's prefix (assigned by the central bank)
        - The last 5 characters are unique within the bank and auto-generated
        - Account numbers are case-insensitive in storage but should be uppercase in responses

        **Currency Support:**
        Branch banks typically support major currencies. The requested currency must be one that the bank supports.

        **Note:** This endpoint requires the user to be authenticated (e.g., via session token or API key).
      operationId: createAccount
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: The unique identifier of the user
          schema:
            type: string
            pattern: '^user-[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
            example: "user-550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreationRequest'
            examples:
              eurosAccount:
                summary: Create Euro account
                value:
                  currency: "EUR"
              dollarsAccount:
                summary: Create US Dollar account
                value:
                  currency: "USD"
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountCreationResponse'
              examples:
                successfulCreation:
                  summary: Euro account created
                  value:
                    accountNumber: "EST12345"
                    ownerId: "user-550e8400-e29b-41d4-a716-446655440000"
                    currency: "EUR"
                    balance: "0.00"
                    createdAt: "2026-04-02T11:15:15Z"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidRequest:
                  summary: Missing or invalid currency
                  value:
                    code: "INVALID_REQUEST"
                    message: "Currency is required and must be a valid ISO 4217 code"
                invalidCurrency:
                  summary: Unsupported currency
                  value:
                    code: "UNSUPPORTED_CURRENCY"
                    message: "Currency 'XXX' is not supported by this bank"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                unauthorized:
                  summary: Authentication required
                  value:
                    code: "UNAUTHORIZED"
                    message: "Authentication is required to create an account"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                userNotFound:
                  summary: User does not exist
                  value:
                    code: "USER_NOT_FOUND"
                    message: "User with ID 'user-550e8400-e29b-41d4-a716-446655440000' not found"
  /accounts/{accountNumber}:
    get:
      tags:
        - Accounts
      summary: Look up account by account number
      description: |
        Retrieves account ownership information for a specific account number.

        **Unauthenticated Access:**
        - This endpoint does NOT require authentication
        - Anyone can look up an account number to verify ownership
        - This enables pre-transfer verification and account validation

        **Response Behavior:**
        - Returns the owner's full name when the account exists
        - Returns 404 Not Found when the account does not exist

        **Use Cases:**
        - Verify account ownership before initiating transfers
        - Validate account numbers provided by users
        - Debug and troubleshooting

        **Privacy Note:**
        Only the owner's name is returned, not their email, balance, or transaction history. This balances
        utility (allowing account verification) with privacy (protecting sensitive financial information).

        **Account Number Format:**
        Account numbers must be exactly 8 characters and match the pattern: `[A-Z0-9]{3}[A-Z0-9]{5}`.
        The first 3 characters identify the bank, and the last 5 are the unique account identifier.
      operationId: lookupAccount
      security: []
      parameters:
        - name: accountNumber
          in: path
          required: true
          description: |
            The 8-character account number to look up.

            Format: `{BANK_PREFIX}{UNIQUE_SUFFIX}`
            - BANK_PREFIX: 3 characters (the bank's identifier)
            - UNIQUE_SUFFIX: 5 characters (unique within the bank)

            Example: "EST12345" where "EST" is the bank prefix
          schema:
            type: string
            pattern: '^[A-Z0-9]{8}$'
            example: "EST12345"
      responses:
        '200':
          description: Account found - owner information returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountLookupResponse'
              examples:
                accountFound:
                  summary: Account exists
                  value:
                    accountNumber: "EST12345"
                    ownerName: "Jane Doe"
                    currency: "EUR"
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                accountNotFound:
                  summary: Account does not exist
                  value:
                    code: "ACCOUNT_NOT_FOUND"
                    message: "Account with number 'EST12345' not found"
        '400':
          description: Invalid account number format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidAccountNumber:
                  summary: Malformed account number
                  value:
                    code: "INVALID_ACCOUNT_NUMBER"
                    message: "Account number must be exactly 8 characters"
...
components:
  schemas:
    UserRegistrationRequest:
      type: object
      required:
        - fullName
      properties:
        fullName:
          type: string
          description: The full name of the user (used in account lookups)
          minLength: 2
          maxLength: 200
          example: "Jane Doe"
        email:
          type: string
          format: email
          description: Optional email address for the user (used for notifications)
          maxLength: 255
          example: "jane.doe@example.com"
    UserRegistrationResponse:
      type: object
      required:
        - userId
        - fullName
        - createdAt
      properties:
        userId:
          type: string
          description: The unique identifier assigned to this user
          pattern: '^user-[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
          example: "user-550e8400-e29b-41d4-a716-446655440000"
        fullName:
          type: string
          description: The full name of the registered user
          example: "Jane Doe"
        email:
          type: string
          format: email
          description: The email address of the user (if provided)
          example: "jane.doe@example.com"
        createdAt:
          type: string
          format: date-time
          description: |
            The timestamp when the user was registered.

            **Note:** This timestamp is stored on the branch bank server and reflects
            the server's clock at the time of registration.
          example: "2026-04-02T11:15:00Z"
    AccountCreationRequest:
      type: object
      required:
        - currency
      properties:
        currency:
          type: string
          description: |
            The ISO 4217 currency code for the account.

            **Supported Currencies:**
            Branch banks typically support major currencies such as EUR, USD, GBP, etc.

            **Validation:**
            The currency must be a valid 3-letter ISO 4217 code and supported by the specific branch bank.

            **Examples:**
            - EUR: Euro
            - USD: United States Dollar
            - GBP: British Pound
          pattern: '^[A-Z]{3}$'
          example: "EUR"
    AccountCreationResponse:
      type: object
      required:
        - accountNumber
        - ownerId
        - currency
        - balance
        - createdAt
      properties:
        accountNumber:
          type: string
          description: |
            The 8-character account number generated for this account.

            **Format:**
            `{BANK_PREFIX}{UNIQUE_SUFFIX}`
            - BANK_PREFIX: 3 characters (matches the bank's central bank assigned prefix)
            - UNIQUE_SUFFIX: 5 characters (unique within this bank, auto-generated)

            **Example:**
            "EST12345" where:
            - "EST" = Estonia Commercial Bank's prefix
            - "12345" = unique identifier within Estonia Commercial Bank

            **Note:** Account numbers are globally unique in the system due to the bank prefix.
          pattern: '^[A-Z0-9]{8}$'
          example: "EST12345"
        ownerId:
          type: string
          description: The unique identifier of the user who owns this account
          pattern: '^user-[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
          example: "user-550e8400-e29b-41d4-a716-446655440000"
        currency:
          type: string
          description: The ISO 4217 currency code for this account
          pattern: '^[A-Z]{3}$'
          example: "EUR"
        balance:
          type: string
          format: decimal
          description: |
            The current balance of the account, as a decimal string.

            **Format:**
            - Stored as a string to avoid floating-point precision issues
            - Always includes 2 decimal places
            - Negative values represent overdrafts (if supported by the bank)

            **Initial Balance:**
            Newly created accounts start with a balance of "0.00".
          pattern: '^-?\d+\.\d{2}$'
          example: "0.00"
        createdAt:
          type: string
          format: date-time
          description: |
            The timestamp when the account was created.

            **Note:** This timestamp is stored on the branch bank server and reflects
            the server's clock at the time of account creation.
          example: "2026-04-02T11:15:15Z"
    AccountLookupResponse:
      type: object
      required:
        - accountNumber
        - ownerName
        - currency
      properties:
        accountNumber:
          type: string
          description: The 8-character account number that was looked up
          pattern: '^[A-Z0-9]{8}$'
          example: "EST12345"
        ownerName:
          type: string
          description: |
            The full name of the account owner.

            **Privacy Note:**
            Only the owner's name is returned. Email addresses, balances, transaction history,
            and other sensitive information are NOT exposed through this unauthenticated endpoint.

            This design allows account verification transfers while protecting user privacy.
          example: "Jane Doe"
        currency:
          type: string
          description: The ISO 4217 currency code for this account
          pattern: '^[A-Z]{3}$'
          example: "EUR"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: A machine-readable error code
          example: "ACCOUNT_NOT_FOUND"
        message:
          type: string
          description: A human-readable error message
          example: "Account with number 'EST12345' not found"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT bearer token for authenticated operations.

        **Required For:**
        - Account creation (POST /users/{userId}/accounts)
        - Transfer initiation (POST /transfers)
        - Transfer status lookup (GET /transfers/{transferId})

        **Not Required For:**
        - User registration (POST /users) - initial signup is open
        - Account lookup (GET /accounts/{accountNumber}) - unauthenticated for verification
        - Inter-bank transfer reception (POST /transfers/receive) - authenticated via bank-signed JWT (D-04)

        **Token Format:**
        Tokens are signed JWTs containing the userId and expiration time.